refactor(codegen): remove context and frame
